Made ImageFromFile work for any standard image resource
The previous code built a path called "Images/" plus the filename.  But
this is a relative path, and Cocoa requires absolute paths.  So I don't
know how this ever worked.

I've left this code in place, but added new code that (in case the old
code doesn't work) does the standard MainBundle.PathForImageResource
dance.  This will find an image in any standard format, as long as it's
added to the application bundle.